+++ CHANGELOG.rst
@@ -1,4 +1,11 @@
 
+Release 0.9.2:
+
+   * Add abigui.py script
+   * Add abilab.abirobot function to create a Robot from a list of filepaths.
+   * Add plotly version of fatbands
+   * Support for dipquad and quadquad
+
 Release 0.9.1:
 
    * Add  "-ew", "--expose-web",